Stufe 5: Anbindung an Mekanism
Was die Standard-Schnittstellen nicht hergeben, holt jetzt eine eigene Bruecke:
Chemikalientanks (seit Mekanism 10.7 ein Typ statt vier), Strom in Joule statt in
umgerechnetem FE, Waermespeicher, der digitale Bergmann und alle sieben
Mehrblockbauten -- Kessel, Induktionsmatrix, Verdunstungsanlage, SPS, Spaltreaktor,
Fusionsanlage, Turbine. Der Waermemelder findet auch einen Mekanism-Spaltreaktor.

Die drei Mekanism-Schnittstellen kommen ueber ihre Namen aus dem Capability-Register
von NeoForge, nicht ueber Mekanisms interne Klasse Capabilities.

Spaltreaktor, Fusionsanlage und Turbine gehoeren zu Mekanism: Generators, einer
eigenen Mod. Zum Uebersetzen braucht die Bruecke beide JARs; im Spiel genuegt
Mekanism selbst, weil MekGenerators nur angesprochen wird, wenn die Mod geladen ist.

Aus dem Muster ist dabei ein Werkzeug geworden: tools/api_check.py bekommt
Verzeichnis, Quellen und Paketvorsaetze von aussen und prueft beide Anbindungen,
build.gradle sucht die JARs ueber eine gemeinsame Funktion (und nicht mehr ueber
"irgendeine JAR in libs/", weil dort jetzt mehrere liegen), und CI baut einmal ohne
und einmal mit allen fremden JARs. Ab dem zweiten Bau tragen alle weiteren Schritte
beide Bruecken -- die hochgeladene JAR ist damit die, die man ausliefern wuerde, und
der Servertest laeuft mit den Bruecken in der JAR, aber ohne die fremden Mods.

Zwei Luecken im Torwaechter sind dabei aufgefallen und geschlossen: X.class galt als
Zugriff auf ein Glied namens "class", und eine Variable, die in derselben Datei
zweimal mit verschiedenem Typ vorkommt, wurde dem zuletzt gesehenen Typ zugerechnet.

+++ src/main/java/com/zuxelus/energycontrol/utils/FluidInfo.java
@@ -11,9 +11,9 @@ import net.neoforged.neoforge.fluids.capability.IFluidHandler;
  * Anbindung an eine Mod baut daraus ihre eigenen Tanks um; der Kern kennt nur diese drei
  * Angaben und die Fluid-Schnittstelle von NeoForge.
  */
-public record FluidInfo(Component name, int amount, int capacity) {
+public record FluidInfo(Component name, long amount, long capacity) {
 
-    public static FluidInfo of(FluidStack stack, int capacity) {
+    public static FluidInfo of(FluidStack stack, long capacity) {
         if(stack == null || stack.isEmpty()) return empty(capacity);
         return new FluidInfo(stack.getHoverName(), stack.getAmount(), capacity);
     }
@@ -22,7 +22,7 @@ public record FluidInfo(Component name, int amount, int capacity) {
         return of(handler.getFluidInTank(tank), handler.getTankCapacity(tank));
     }
 
-    public static FluidInfo empty(int capacity) {
+    public static FluidInfo empty(long capacity) {
         return new FluidInfo(Component.translatable("msg.ec.InfoPanelEmpty"), 0, capacity);
     }
